+++ config.js
@@ -1,12 +1,13 @@
 // Client side things only, don't put things here that shouldn't be public
 export const GRAPHQL_ENDPOINT = `https://strapi.christopherleemiller.me/graphql`;
 export const STRAPI_ENDPOINT = `https://strapi.christopherleemiller.me`;
-export const CLOUDINARY_URL = `https://res.cloudinary.com/christopherleemiller/image/upload`;
-export const CLOUDINARY_CLOUD = `v1554111995`;
+export const CLOUDINARY_URL = `https://res.cloudinary.com`;
+export const CLOUDINARY_CLOUD = `christopherleemiller`;
+export const CLOUDINARY_VERSION = `v1554111995`;
 export const CLOUDINARY_FOLDER = `clm_me`;
 
 export const SITE_TITLE = `Christopher Lee Miller`;
-export const SITE_DEFAULT_IMAGE = `${CLOUDINARY_URL}/c_scale,w_300/${CLOUDINARY_CLOUD}/${CLOUDINARY_FOLDER}/assets/logo.png`;
+export const SITE_DEFAULT_IMAGE = `${CLOUDINARY_URL}/${CLOUDINARY_CLOUD}/image/upload/c_scale,w_300/${CLOUDINARY_VERSION}/${CLOUDINARY_FOLDER}/assets/logo.png`;
 export const SEPARATOR = ` - `;
 
 export const DEFAULT_SEO = {
